+++ sdk/@launchdarkly/observability/src/index.ts
@@ -1,1 +1,2 @@
 export { Observe as default, LDObserve } from 'highlight.run'
+export type { ObserveOptions } from 'highlight.run'