+++ sdk/@launchdarkly/session-replay/src/index.ts
@@ -1,1 +1,2 @@
 export { Record as default, LDRecord } from 'highlight.run'
+export type { RecordOptions } from 'highlight.run'