+++ sdk/highlight-run/src/client/index.tsx
@@ -404,13 +404,7 @@ export class Highlight {
 		const client = new GraphQLClient(`${this._backendUrl}`, {
 			headers: {},
 		})
-		this.graphqlSDK = getSdk(
-			client,
-			getGraphQLRequestWrapper(
-				this.sessionData?.sessionSecureID ||
-					this.options?.sessionSecureID,
-			),
-		)
+		this.graphqlSDK = getSdk(client, getGraphQLRequestWrapper())
 		this.environment = options.environment ?? 'production'
 		this.appVersion = options.appVersion
 		this.serviceName = options.serviceName ?? ''