+++ sdk/highlight-run/src/client/workers/highlight-client-worker.ts
@@ -373,7 +373,7 @@ function stringifyProperties(
 				new GraphQLClient(backend, {
 					headers: {},
 				}),
-				getGraphQLRequestWrapper(sessionSecureID),
+				getGraphQLRequestWrapper(),
 			)
 			return
 		}